// src/app/core/demofile-reader/models/player-round-stats-model.ts
export interface PlayerRoundStatsModel {

    kills: number;

    damage: number;

    equipmentValue: number;

    moneySaved: number;

    killReward: number;

    liveTime: number;

    deaths: number;

    assists: number;

    headShotKills: number;
    
    objective: number;

    enemiesFlashed: number;

    utilityDamage: number;
    
}